Added undo/redo, improved navigation, and selecting suggestions
Adding Ctrl + Z and Ctrl + Y functionality for when input is focused. Actions that are added/removed stack of actions which can be undone and added to another stack of actions, which can be redone.

Added focusing after render after making a selection.

Now after a suggestion is chosen, the largest amount of matching search text is removed, and any remaining text stays in the search input