Added a callback for first bean load
After beans are first loaded with the widest possible filters, it is then passed back to Beans and then used to generate auto-complete options so no options that result in no beans can be show. E.g. if a roaster hasn't given permission yet, their roaster name will not show as an autocomplete option